requirements.txt
Adafruit-Blinka==4.6.0
adafruit-circuitpython-dht==3.4.1
Adafruit-PlatformDetect==2.8.0
Adafruit-PureIO==1.1.5
alembic==1.4.2
astroid==2.3.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
pyftdi==0.51.0
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pyusb==1.0.2
requests==2.24.0
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
six==1.14.0
SQLAlchemy==1.3.16
sysv-ipc==1.0.1
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
